rename/reorganize the folders to have a better package structure
