+++ .github/workflows/pr_checks.yml
@@ -5,71 +5,148 @@ on:
     branches:
       - main
 
+env:
+  CONTAINER: fideslib-local
+  IMAGE: fideslib:local
+
 jobs:
-  Black:
+  Build:
     runs-on: ubuntu-latest
     steps:
+      - name: Checkout
+        uses: actions/checkout@v2
+
+      - name: Set up Docker Buildx
+        id: buildx
+        uses: docker/setup-buildx-action@v1
+
+      - name: Build fidesctl container
+        uses: docker/build-push-action@v2
+        with:
+          builder: ${{ steps.buildx.outputs.name }}
+          context: ./fidesctl
+          file: ./fidesctl/Dockerfile
+          outputs: type=docker,dest=/tmp/${{ env.CONTAINER }}.tar
+          push: false
+          tags: ${{ env.IMAGE }}
+
+      - name: Upload fidesctl container
+        uses: actions/upload-artifact@v2
+        with:
+          name: ${{ env.CONTAINER }}
+          path: /tmp/${{ env.CONTAINER }}.tar
+          retention-days: 1
+
+  Fidesctl:
+    needs: Build
+    runs-on: ubuntu-latest
+    steps:
+      - name: Download fidesctl container
+        uses: actions/download-artifact@v2
+        with:
+          name: ${{ env.CONTAINER }}
+          path: /tmp/
+
+      - name: Load fidesctl image
+        run: docker load --input /tmp/${{ env.CONTAINER }}.tar
+
+      - name: Checkout
+        uses: actions/checkout@v2
+
+      - name: Check package installation
+        run: make check-install
+
+  Black:
+    needs: Build
+    runs-on: ubuntu-latest
+    steps:
+      - name: Download fidesctl container
+        uses: actions/download-artifact@v2
+        with:
+          name: ${{ env.CONTAINER }}
+          path: /tmp/
+
+      - name: Load fidesctl image
+        run: docker load --input /tmp/${{ env.CONTAINER }}.tar
+
       - name: Checkout
         uses: actions/checkout@v2
 
       - name: Run formatter
-        uses: psf/black@stable
-        with:
-          options: "--check --verbose"
-          version: "21.8b0"
+        run: make black
 
   Pylint:
+    needs: Build
     runs-on: ubuntu-latest
     steps:
+      - name: Download fidesctl container
+        uses: actions/download-artifact@v2
+        with:
+          name: ${{ env.CONTAINER }}
+          path: /tmp/
+
+      - name: Load fidesctl image
+        run: docker load --input /tmp/${{ env.CONTAINER }}.tar
+
       - name: Checkout
         uses: actions/checkout@v2
-
-      - name: Install dev dependencies
-        run: |
-          python -m pip install --upgrade pip
-          if [ -f dev-requirements.txt ]; then pip install -r dev-requirements.txt; fi
 
       - name: Run linter
-        run: pylint --jobs=0 ./**
+        run: make pylint
 
   Mypy:
+    needs: Build
     runs-on: ubuntu-latest
     steps:
+      - name: Download fidesctl container
+        uses: actions/download-artifact@v2
+        with:
+          name: ${{ env.CONTAINER }}
+          path: /tmp/
+
+      - name: Load fidesctl image
+        run: docker load --input /tmp/${{ env.CONTAINER }}.tar
+
       - name: Checkout
         uses: actions/checkout@v2
-
-      - name: Install dev dependencies
-        run: |
-          python -m pip install --upgrade pip
-          if [ -f dev-requirements.txt ]; then pip install -r dev-requirements.txt; fi
 
       - name: Run typechecker
-        run: mypy --exclude 'setup\.py$' --ignore-missing-imports ./
+        run: make mypy
 
   Xenon:
+    needs: Build
     runs-on: ubuntu-latest
     steps:
+      - name: Download fidesctl container
+        uses: actions/download-artifact@v2
+        with:
+          name: ${{ env.CONTAINER }}
+          path: /tmp/
+
+      - name: Load fidesctl image
+        run: docker load --input /tmp/${{ env.CONTAINER }}.tar
+
       - name: Checkout
         uses: actions/checkout@v2
-
-      - name: Install dev dependencies
-        run: |
-          python -m pip install --upgrade pip
-          if [ -f dev-requirements.txt ]; then pip install -r dev-requirements.txt; fi
 
       - name: Run cyclomatic complexity check
-        run: xenon ./ --max-absolute=B -max-modules=B --max-average=A
+        run: make xenon
 
   Pytest:
+    needs: Build
     runs-on: ubuntu-latest
     steps:
+      - name: Download fidesctl container
+        uses: actions/download-artifact@v2
+        with:
+          name: ${{ env.CONTAINER }}
+          path: /tmp/
+
+      - name: Load fidesctl image
+        run: docker load --input /tmp/${{ env.CONTAINER }}.tar
+
       - name: Checkout
         uses: actions/checkout@v2
 
-      - name: Install dev dependencies
-        run: |
-          python -m pip install --upgrade pip
-          if [ -f dev-requirements.txt ]; then pip install -r dev-requirements.txt; fi
-
-      - name: Run unit test suite
-        run: pytest
+      - name: Run test suite
+        run: make pytest
