+++ dev-requirements.txt
@@ -1,4 +1,5 @@
 black==21.8b0
+ipython==8.0.1
 mypy==0.910
 pylint==2.6.0
 pytest==6.2.2
