+++ setup.py
@@ -2,25 +2,26 @@ from pathlib import Path
 from setuptools import setup, find_packages
 from typing import List
 
-# Packages
-CRYPTO = "cryptography"
-OAUTH = "oauth"
+
+def read_requirements(filename: str) -> List[str]:
+    with open(filename, "r") as file:
+        requirements = file.read().strip().split("\n")
+    return requirements
 
 
-def parse_local_file(filename: str) -> str:
-    here = Path(__file__).parent.resolve()
-    return (here / filename).read_text("utf-8").strip()
+# Requirements and Dependencies
+dev_requires = read_requirements("dev-requirements.txt")
+cryptography_requires = read_requirements("fideslib/cryptography/requirements.txt")
+oauth_requires = read_requirements("fideslib/oauth/requirements.txt")
 
-
-def list_requirements(filename: str) -> List[str]:
-    return parse_local_file(filename).split("\n")
-
+extras = {"crypto": cryptography_requires, "oauth": oauth_requires}
+extras["all"] = sum([value for key, value in extras.items()], [])
 
 setup(
     name="fideslib",
     version="0.0.1",
     description="Shared libraries, for use in any fides project.",
-    long_description=parse_local_file("README.md"),
+    long_description=open("README.md").read(),
     long_description_content_type="text/markdown",
     url="https://github.com/ethyca/fideslib",
     author="Ethyca, Inc.",
@@ -34,15 +35,8 @@ setup(
         "Programming Language :: Python :: 3.8",
         "Topic :: Software Development :: Libraries",
     ],
-    package_dir={
-        CRYPTO: CRYPTO,
-        OAUTH: OAUTH,
-    },
-    packages=find_packages(where="src"),
+    package_dir={"": "fideslib"},
     python_requires=">=3.8, <4",
-    dev_requires=list_requirements("dev-requirements.txt"),
-    extras_require={
-        "crypto": list_requirements(f"src/{CRYPTO}/requirements.txt"),
-        "oauth": list_requirements(f"src/{OAUTH}/requirements.txt"),
-    },
+    dev_requires=dev_requires,
+    extras_require=extras,
 )
